Quote Visualization

takes a dataframe and makes a gt table or ggplot that shows a quote

quote_viz(
  data,
  text_col = colnames(data)[1],
  viz_type = "gt",
  custom_highlight = NULL,
  n = 3,
  print = T,
  width = 60,
  title = NULL,
  suppress_warnings = T,
  align = "center",
  save = T,
  ...
)

Arguments

data

the dataframe

text_col

columns to create table for

viz_type

ggplot or gt visualization

custom_highlight

a vector, optional custom highlighting

n

integer, number of words to auto-highlight

print

T, whether or not to print the highlighted words to console

width

The width of the table generated

title

the title of the ggplot or gt

suppress_warnings

T/F suppression of warnings

align

the table alignment: "left", "center", "right"

save

if TRUE actually make the gt

...

Arguments passed onto the gt table

Value

a ggplot/gt that visualizes text

Examples

if (FALSE) {
df <- TeachingLab::survey_monkey
colnames(df)[1] <- "What learning are you excited to try?"
quote_viz(
  data = df,
  text_col = "What learning are you excited to try?",
  viz_type = "gt",
  title = "Responses from Survey Monkey"
)
}