Teaching Lab Custom Ggplot2 Theme

It requires installing Roboto, Calibri fonts unless you change the font parameters

https://www.google.com/fonts

theme_tl(
  base_family = "Calibri",
  base_size = 14,
  strip_text_family = base_family,
  strip_text_size = 15,
  plot_title_family = "Calibri",
  plot_title_size = 20,
  plot_title_margin = 10,
  subtitle_family = "Roboto",
  subtitle_size = 15,
  subtitle_margin = 15,
  caption_family = "Roboto",
  caption_size = 11,
  caption_margin = 10,
  axis_title_family = "Calibri",
  axis_title_size = 12,
  axis_title_just = "mm",
  axis_text_size = 10.5,
  dark = FALSE,
  grid = TRUE,
  axis = FALSE,
  ticks = FALSE,
  markdown = FALSE,
  legend = F
)

Arguments

base_family

base font family

base_size

base font size

strip_text_family

facet label font family

strip_text_size

facet label text size

plot_title_family

plot title family

plot_title_size

plot title font size

plot_title_margin

plot title margin

subtitle_family

plot subtitle family

subtitle_size

plot subtitle size

subtitle_margin

plot subtitle margin

caption_family

plot caption family

caption_size

plot caption size

caption_margin

plot caption margin

axis_title_family

axis title font family

axis_title_size

axis title font size

axis_title_just

axis title font justification blmcrt

axis_text_size

axis text size

dark

dark mode TRUE, FALSE

grid

panel grid (TRUE, FALSE, or a combination of X, x, Y, y)

axis

axis TRUE, FALSE

ticks

ticks TRUE, FALSE

markdown

enabled ggtext markdown styling TRUE, FALSE

legend

default no legend with F